[Feat] product

+++ src/main/java/TRaMis8khae/starbucks/product/entity/Product.java
@@ -26,7 +26,7 @@ public class Product extends BaseEntity {
     @Column(nullable = false, length = 50)
     private String productUUID;
 
-    private Integer productScore = 0;
+    private Double productScore = 0.0;
 
     @Column(nullable = false)
     private Double price;
